+++ include/Board.h
@@ -1,9 +1,9 @@
 #ifndef __BOARD_H__
 #define __BOARD_H__
 
-#include "boardPosition.h"
-#include "boardSquare.h"
-#include "boardRepresentation.h"
+#include <BoardPosition.h>
+#include <BoardSquare.h>
+#include <BoardRepresentation.h>
 #include <map>
 #include <vector>
