+++ include/Bonus.h
@@ -1,7 +1,7 @@
 #ifndef __BONUS_H__
 #define __BONUS_H__
 
-#include "boardRepresentation.h"
+#include <BoardRepresentation.h>
 
 class Bonus {
