+++ include/Model3D.h
@@ -1,0 +1,18 @@
+#ifndef __MODEL_3D_H__
+#define __MODEL_3D_H__
+
+#include <GL/glew.h>
+
+class Model3D {
+
+public:
+	
+	Model3D(); //TODO should be private:load from file
+
+private:
+	
+	GLuint vbo;
+
+};
+
+#endif
