+++ include/PacGomme.h
@@ -1,9 +1,9 @@
 #ifndef __PAC_GOMME_H__
 #define __PAC_GOMME_H__
 
-#include "pacman.h"
-#include "bonus.h"
-#include "boardRepresentation.h"
+#include <Pacman.h>
+#include <Bonus.h>
+#include <BoardRepresentation.h>
 
 class PacGomme : public Bonus {
