+++ include/Renderer.h
@@ -1,7 +1,7 @@
 #ifndef __RENDERER_H__
 #define __RENDERER_H__
 
-#include "boardRepresentation.h"
+#include <BoardRepresentation.h>
 
 class Renderer {
