+++ src/Model3D.cpp
@@ -1,0 +1,10 @@
+#include <Model3D.h>
+
+#include <GL/glew.h>
+
+Model3D::Model3D() {
+	glGenBuffers(1, &vbo);
+	glBindBuffer(GL_ARRAY_BUFFER, vbo);
+		//TODO glBufferData
+	glBindBuffer(GL_ARRAY_BUFFER, 0);
+}
